# server/tests/fixtures/v1-openapi.json
{
  "components": {
    "schemas": {
...
      "app.announcement.GetListResponse": {
        "additionalProperties": false,
        "properties": {
          "announcements": {
            "items": {
              "$ref": "#/components/schemas/entity.Announcement"
            },
            "type": "array"
          },
          "externalContents": {
            "items": {
              "$ref": "#/components/schemas/entity.AnnouncementExternalContent"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "app.announcement.GetResponse": {
        "additionalProperties": false,
        "properties": {
          "announcement": {
            "$ref": "#/components/schemas/entity.Announcement"
          }
        },
        "type": "object"
      },
...
      "app.arena.DeckTrendResponse": {
        "additionalProperties": false,
        "properties": {
          "leaders": {
            "items": {
              "$ref": "#/components/schemas/entity.ArenaDeckTrendMemberCard"
            },
            "type": "array"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/entity.ArenaDeckTrendMemberCard"
            },
            "type": "array"
          },
          "supports": {
            "items": {
              "$ref": "#/components/schemas/entity.ArenaDeckTrendSupportCard"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "entity.Announcement": {
        "additionalProperties": false,
        "properties": {
          "bannerPath": {
            "type": "string"
          },
          "bannerUrl": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "category": {
            "oneOf": [
              {
                "enum": [
                  "UNSPECIFIED",
                  "OTHER",
                  "EVENT",
                  "GACHA",
                  "CAMPAIGN",
                  "UPDATE",
                  "MAINTENANCE",
                  "BUG"
                ],
                "type": "string"
              },
              {
                "format": "int32",
                "type": "integer"
              }
            ]
          },
          "endAt": {
            "pattern": "^-?[0-9]+$",
            "type": "string"
          },
          "id": {
            "pattern": "^-?[0-9]+$",
            "type": "string"
          },
          "lastUpdatedAt": {
            "pattern": "^-?[0-9]+$",
            "type": "string"
          },
          "pickupBannerPath": {
            "type": "string"
          },
          "pickupBannerUrl": {
            "type": "string"
          },
          "pickupEndAt": {
            "pattern": "^-?[0-9]+$",
            "type": "string"
          },
          "pickupStartAt": {
            "pattern": "^-?[0-9]+$",
            "type": "string"
          },
          "sortOrder": {
            "type": "integer"
          },
          "startAt": {
            "pattern": "^-?[0-9]+$",
            "type": "string"
          },
          "stylesheetId": {
            "pattern": "^-?[0-9]+$",
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/announcement": {
      "get": {
        "description": "No request body. Public mode removes unapproved/account-specific fields; raw mode is operator-only. Query string limited to 8192 bytes and 256 parameters.",
        "operationId": "announcement",
        "parameters": [
          {
            "deprecated": false,
            "description": "Single occurrence only. Nested filters use dotted field names.",
            "explode": true,
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "pattern": "^-?[0-9]+$",
              "type": "string"
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app.announcement.GetResponse"
                }
              }
            },
            "description": "Raw protobuf JSON; 64-bit integers are strings. Fetch time is Unix milliseconds.",
            "headers": {
              "X-Moenotes-Cache": {
                "schema": {
                  "enum": [
                    "HIT",
                    "MISS",
                    "COALESCED"
                  ],
                  "type": "string"
                }
              },
              "X-Moenotes-Fetched-At": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query"
          },
          "401": {
            "description": "Missing or invalid HTTP API key"
          },
          "403": {
            "description": "Route prohibited by response policy"
          },
          "404": {
            "description": "Route disabled or unknown"
          },
          "405": {
            "description": "Only GET is supported; HEAD does not query upstream"
          },
          "429": {
            "description": "Local request queue full"
          },
          "502": {
            "description": "Upstream business, transport or protocol error"
          },
          "503": {
            "description": "Upstream session or availability blocked"
          },
          "504": {
            "description": "Request deadline exceeded"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "tags": [
          "read queries"
        ]
      }
    },
    "/v1/announcements": {
      "get": {
        "description": "No request body. Public mode removes unapproved/account-specific fields; raw mode is operator-only. Query string limited to 8192 bytes and 256 parameters.",
        "operationId": "announcements",
        "parameters": [
          {
            "deprecated": false,
            "description": "Single occurrence only. Nested filters use dotted field names.",
            "explode": true,
            "in": "query",
            "name": "selectedTab",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "enum": [
                    "ALL",
                    "OPENING_EVENTS",
                    "BUG"
                  ],
                  "type": "string"
                },
                {
                  "format": "int32",
                  "type": "integer"
                }
              ]
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app.announcement.GetListResponse"
                }
              }
            },
            "description": "Raw protobuf JSON; 64-bit integers are strings. Fetch time is Unix milliseconds.",
            "headers": {
              "X-Moenotes-Cache": {
                "schema": {
                  "enum": [
                    "HIT",
                    "MISS",
                    "COALESCED"
                  ],
                  "type": "string"
                }
              },
              "X-Moenotes-Fetched-At": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query"
          },
          "401": {
            "description": "Missing or invalid HTTP API key"
          },
          "403": {
            "description": "Route prohibited by response policy"
          },
          "404": {
            "description": "Route disabled or unknown"
          },
          "405": {
            "description": "Only GET is supported; HEAD does not query upstream"
          },
          "429": {
            "description": "Local request queue full"
          },
          "502": {
            "description": "Upstream business, transport or protocol error"
          },
          "503": {
            "description": "Upstream session or availability blocked"
          },
          "504": {
            "description": "Request deadline exceeded"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "tags": [
          "read queries"
        ]
      }
    },
    "/v1/arena/deck-trend": {
      "get": {
        "description": "No request body. Public mode removes unapproved/account-specific fields; raw mode is operator-only. Query string limited to 8192 bytes and 256 parameters.",
        "operationId": "deck-trend",
        "parameters": [
          {
            "deprecated": false,
            "description": "Single occurrence only. Nested filters use dotted field names.",
            "explode": true,
            "in": "query",
            "name": "musicId",
            "required": true,
            "schema": {
              "pattern": "^-?[0-9]+$",
              "type": "string"
            },
            "style": "form"
          },
          {
            "deprecated": false,
            "description": "Single occurrence only. Nested filters use dotted field names.",
            "explode": true,
            "in": "query",
            "name": "arenaSeasonId",
            "required": true,
            "schema": {
              "pattern": "^-?[0-9]+$",
              "type": "string"
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app.arena.DeckTrendResponse"
                }
              }
            },
            "description": "Raw protobuf JSON; 64-bit integers are strings. Fetch time is Unix milliseconds.",
            "headers": {
              "X-Moenotes-Cache": {
                "schema": {
                  "enum": [
                    "HIT",
                    "MISS",
                    "COALESCED"
                  ],
                  "type": "string"
                }
              },
              "X-Moenotes-Fetched-At": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query"
          },
          "401": {
            "description": "Missing or invalid HTTP API key"
          },
          "403": {
            "description": "Route prohibited by response policy"
          },
          "404": {
            "description": "Route disabled or unknown"
          },
          "405": {
            "description": "Only GET is supported; HEAD does not query upstream"
          },
          "429": {
            "description": "Local request queue full"
          },
          "502": {
            "description": "Upstream business, transport or protocol error"
          },
          "503": {
            "description": "Upstream session or availability blocked"
          },
          "504": {
            "description": "Request deadline exceeded"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "tags": [
          "read queries"
        ]
      }
    },
...
    "/v1/status": {
      "get": {
        "responses": {
          "200": {
            "description": "Sanitized local operational state"
          },
          "401": {
            "description": "Missing API key"
          },
          "503": {
            "description": "Not ready"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    }
  }
}